메인페이지 레이아웃 수정
