+++ frontend/src/common/component/recharts/MonthlyProductSalesPieChart.scss
@@ -1,4 +1,4 @@
-@import "../App";
+@import "../../../App";
 
 .select-container {
   display: flex;
