+++ frontend/src/common/component/recharts/QuaterProfitGaugeChart.js
